requirements.txt
# Core dependencies for Trestle ETL Pipeline
requests>=2.31.0
pymysql>=1.1.0
python-dotenv>=1.0.0

# Testing dependencies
hypothesis>=6.92.0
pytest>=7.4.0
pytest-cov>=4.1.0
